+++ maps.earth/Views/PlaceField.swift
@@ -13,12 +13,12 @@ private let logger = FileLogger()
 struct PlaceField: View {
   var header: String
   @Binding var place: Place?
-  @State var isSearching: Bool = false
+  @State var searchIsPresented: Bool = false
   @State var queryText: String = ""
   @StateObject var searchQueue: SearchQueue = SearchQueue()
 
   var body: some View {
-    Button(action: { isSearching = true }) {
+    Button(action: { searchIsPresented = true }) {
       HStack(spacing: 16) {
         Text("\(header):").foregroundColor(.hw_darkGray)
           // this minWidth is intended to approximately align the To/From heafer
@@ -31,62 +31,29 @@ struct PlaceField: View {
     }.onChange(of: queryText) { newValue in
       searchQueue.textDidChange(newValue: newValue)
     }
-    .sheet(isPresented: $isSearching) {
-      PlaceSearch(
-        placeholder: header,
-        hasPendingQuery: searchQueue.hasPendingQuery,
-        places: searchQueue.mostRecentResults,
-        selectedPlace: $place
-      ).searchable(text: $queryText)
-    }
-  }
-}
-
-struct PlaceSearch: View {
-  var placeholder: String
-  var hasPendingQuery: Bool
-  var places: [Place]?
-  @Binding var selectedPlace: Place?
-
-  @Environment(\.isSearching) private var isSearching
-  @Environment(\.dismissSearch) private var dismissSearch
-  @Environment(\.dismiss) private var dismiss
-
-  var body: some View {
-    NavigationView {
-      VStack {
-        if hasPendingQuery {
-          Text("Looking... 🧐")
-        }
-        if let places = places {
-          if places.isEmpty && !hasPendingQuery {
-            Text("No results. 😢")
-          }
-          List(places, selection: $selectedPlace) { place in
-            PlaceRow(place: place).onTapGesture {
-              print("selected a place, dismissing search")
-              selectedPlace = place
-              dismissSearch()
-              dismiss()
+    .sheet(isPresented: $searchIsPresented) {
+      SheetContents(
+        title: "Change Stop", onClose: { searchIsPresented = false },
+        currentDetent: .constant(.large)
+      ) {
+        ScrollView {
+          PlaceSearch(
+            placeholder: header,
+            hasPendingQuery: searchQueue.hasPendingQuery,
+            places: $searchQueue.mostRecentResults,
+            queryText: $queryText,
+            didDismissSearch: {
+              searchIsPresented = false
+            },
+            didSubmitSearch: {
+              searchIsPresented = false
+            },
+            didSelectPlace: { newPlace in
+              searchIsPresented = false
+              place = newPlace
             }
-          }.hwListStyle()
+          )
         }
-        Spacer()
-      }
-      // TODO: This should be "sheetColor" but it doesnt play nice with the built in
-      // search controller
-      .background(Color.white)
-      .navigationTitle("Change Stop")
-      .toolbar {
-        Button(action: { dismiss() }) {
-          Image(systemName: "xmark")
-        }
-      }
-    }
-    .onChange(of: isSearching) { newValue in
-      if !newValue {
-        dismissSearch()
-        dismiss()
       }
     }
   }
@@ -106,5 +73,5 @@ struct PlaceSearch: View {
 #Preview("searching place field") {
   let searchQueue = SearchQueue(mostRecentResults: FixtureData.places.all)
   return PlaceField(
-    header: "From", place: .constant(nil), isSearching: true, searchQueue: searchQueue)
+    header: "From", place: .constant(nil), searchIsPresented: true, searchQueue: searchQueue)
 }
