+++ maps.earth/Views/PlaceList.swift
@@ -14,27 +14,15 @@ protocol PlaceListDelegate {
 
 struct PlaceList: View {
   @Binding var places: [Place]?
-  @Binding var selectedPlace: Place?
-  @ObservedObject var tripPlan: TripPlan
-  @Binding var placeDetailsDetent: PresentationDetent
+  var didSelectPlace: (Place) -> Void
 
   var body: some View {
-    let hasSelectedPlace = Binding(
-      get: { selectedPlace != nil },
-      set: { newValue in
-        if newValue {
-          assert(selectedPlace != nil)
-        } else {
-          selectedPlace = nil
-        }
-      })
-
     if let places = places {
       VStack {
         ForEach(places) { place in
           PlaceRow(place: place).onTapGesture {
             dismissKeyboard()
-            selectedPlace = place
+            didSelectPlace(place)
           }
           .padding(.vertical, 4)
           Divider()
@@ -46,26 +34,6 @@ struct PlaceList: View {
       // first responder here - and it's something that the sheet logic
       // reacts to, not just the keyboard.
       // .scrollDismissesKeyboard(.immediately)
-
-      .sheet(isPresented: hasSelectedPlace) {
-        if let selectedPlace = selectedPlace {
-          SheetContents(
-            title: selectedPlace.name,
-            onClose: {
-              self.selectedPlace = nil
-            },
-            currentDetent: $placeDetailsDetent
-          ) {
-            ScrollView {
-              PlaceDetail(place: selectedPlace, tripPlan: tripPlan)
-            }
-            // This is arguably useful.
-            // Usually I just want to swipe down to get a better look at the map without closing out
-            // of the place. If I actually want to dismiss, it's easy enough to hit the X
-            .interactiveDismissDisabled(true)
-          }
-        }
-      }
     } else {
       Text("Loading...")
     }
@@ -74,14 +42,12 @@ struct PlaceList: View {
 
 #Preview("inital") {
   PlaceList(
-    places: .constant(FixtureData.places.all), selectedPlace: .constant(nil), tripPlan: TripPlan(),
-    placeDetailsDetent: .constant(.medium))
+    places: .constant(FixtureData.places.all),
+    didSelectPlace: { _ in })
 }
 
 #Preview("selected") {
   PlaceList(
     places: .constant(FixtureData.places.all),
-    selectedPlace: .constant(FixtureData.places[.zeitgeist]),
-    tripPlan: TripPlan(),
-    placeDetailsDetent: .constant(.medium))
+    didSelectPlace: { _ in })
 }
