+++ maps.earthTests/PreferencesControllerTests.swift
@@ -52,7 +52,7 @@ final class PreferencesControllerTests: XCTestCase {
     await fulfillment(of: [expectation], timeout: 10)
   }
 
-  func testOnly5() async {
+  func testOnly10() async {
     let expectation = XCTestExpectation(description: "Open a file asynchronously.")
     Task { @MainActor in
       let preferencesController = PreferencesController(
@@ -64,11 +64,16 @@ final class PreferencesControllerTests: XCTestCase {
       try! await preferencesController.addSearch(text: "three")
       try! await preferencesController.addSearch(text: "four")
       try! await preferencesController.addSearch(text: "five")
-      XCTAssertEqual(
-        preferencesController.preferences.recentSearches, ["five", "four", "three", "two", "one"])
       try! await preferencesController.addSearch(text: "six")
+      try! await preferencesController.addSearch(text: "seven")
+      try! await preferencesController.addSearch(text: "eight")
+      try! await preferencesController.addSearch(text: "nine")
+      try! await preferencesController.addSearch(text: "ten")
       XCTAssertEqual(
-        preferencesController.preferences.recentSearches, ["six", "five", "four", "three", "two"])
+        preferencesController.preferences.recentSearches, ["ten", "nine", "eight", "seven", "six", "five", "four", "three", "two", "one"])
+      try! await preferencesController.addSearch(text: "eleven")
+      XCTAssertEqual(
+      preferencesController.preferences.recentSearches, ["eleven", "ten", "nine", "eight", "seven", "six", "five", "four", "three", "two"])
       expectation.fulfill()
     }
     await fulfillment(of: [expectation], timeout: 10)
